requirements.txt
mariadb
python-dotenv
boto3
